# Conflicts:
                                      #	src/main/java/com/example/usrweb/controller/BookController.java
                                      #	src/main/resources/application.yml

+++ src/main/java/com/example/usrweb/entity/mapEntiry/UserHasCopyright.java
@@ -36,5 +36,5 @@ public class UserHasCopyright extends AbstractEntity<UserHasCopyright> {
 	@ApiModelProperty(name = "userId" , value = "用户id")
 	private String userId;
 	@ApiModelProperty(name = "copyrightId" , value = "著作权id")
-	private Integer copyrightId;
+	private Long copyrightId;
 }
